# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateTagRequest:
    properties:
      name:
        type: string
    type: object
  requests.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category:
        type: string
      category_id:
        type: string
      currency:
        type: string
      date:
        type: string
      date_format:
        type: string
      description:
        type: string
      transaction_type:
        type: string
      transaction_type_id:
        type: string
    type: object
...
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  responses.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  responses.TransactionResponse:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      date:
        type: integer
      description:
        type: string
      id:
        type: integer
      payee:
        type: string
      reference:
        type: string
      transaction_status:
        type: string
      transaction_type:
        type: string
    type: object
  responses.TransactionSchema:
    properties:
      computations:
        items:
          properties:
            formula:
              type: string
            name:
              type: string
          type: object
        type: array
      date_format:
        type: string
      mapping:
        items:
          properties:
            column:
              type: string
            default: {}
            name:
              type: string
            type:
              type: string
          type: object
        type: array
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Haven Finance API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Haven Finance API
  version: "1.0"
paths:
...
  /transactions:
    get:
      description: Retrieve all transactions
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionResponse'
            type: array
      security:
      - AuthToken: []
      summary: Get all transactions
      tags:
      - transactions
  /transactions/{id}:
    get:
      description: Retrieve a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TransactionResponse'
      security:
      - AuthToken: []
      summary: Get a transaction
      tags:
      - transactions
  /transactions/{id}/tags:
    get:
      description: Retrieve all tags for a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TagResponse'
            type: array
      security:
      - AuthToken: []
      summary: Get all tags for a transaction
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Add a tag to a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Create Tag Request
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTagRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - AuthToken: []
      summary: Add a tag to a transaction
      tags:
      - transactions
  /transactions/create:
    post:
      consumes:
      - application/json
      description: Create a transaction
      parameters:
      - description: Create Transaction Request
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTransactionRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - AuthToken: []
      summary: Create a transaction
      tags:
      - transactions
  /transactions/schemas:
    get:
      description: Retrieve transaction schemas
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TransactionSchema'
            type: array
      security:
      - AuthToken: []
      summary: Get transaction schemas
      tags:
      - transactions
...
swagger: "2.0"